feat(projects): allow creating project from meeting dialog
Implements a user flow where a new project can be created directly
from the 'Add Meeting' dialog.

When a user selects '+ Add new project...', the meeting dialog closes
and the project creation dialog opens. After the project is created,
the user is returned to the meeting form, where the new project is
now available and selected. This improves the workflow significantly.